feat: multiphase LBM advances — coupled free-surface, analytical gas, CG collision, bluff-body benchmarks
Verified results:
- Coupled liquid-gas free surface (Körner+BGK): dam break p_gas=0.30, ship FS Ct 72-93%
- Analytical gas compression/expansion: ratio=1.000, RP equation validation
- IBM bubble rise: cy=115.7 (liquid-side physics correct)
- CG advanced collision: Cumulant/Cascaded/KBC D3Q19 (stable, 25ms/step)
- Bluff-body benchmark matrix: Ahmed/soccer/CAARC/Silsoe/tube bundle
- Propeller open-water D3Q27: KT=0.112
- Entropy-based MRT, IBM vectorized, wall pressure, Allen-Cahn framework
- General simulation framework, hull free-surface v2
- Frontend/backend: cylinder interactive, simulations router, i18n

7 new src files, 13 new examples, 45 modified files

+++ src/tensorlbm/d2q9.py
@@ -49,10 +49,10 @@ def equilibrium(
     if device is None:
         device = rho.device
     c = _c_on(device)
-    w = _w_on(device).view(9, 1, 1)
+    w = _w_on(device).to(rho.dtype).view(9, 1, 1)
 
     u_sq = ux * ux + uy * uy
-    cu = c[:, 0].view(9, 1, 1) * ux + c[:, 1].view(9, 1, 1) * uy
+    cu = c[:, 0].view(9, 1, 1).to(rho.dtype) * ux + c[:, 1].view(9, 1, 1).to(rho.dtype) * uy
     return w * rho.unsqueeze(0) * (1.0 + 3.0 * cu + 4.5 * cu * cu - 1.5 * u_sq.unsqueeze(0))
